[enclave/sidechain] cleanup substratee-worker-primitives
