+++ primitives/worker/Cargo.toml
@@ -16,8 +16,9 @@ chrono          = { version = "0.4.19", default-features = false, features = ["a
 substratee-storage = { path = "../storage", default-features = false }
 
 # substrate-deps
-sp-runtime = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master"}
 sp-core = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master"}
+sp-std = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master"}
+sp-runtime = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master"}
 
 [features]
 default = ["std"]
